int geonames

+++ newsapi.py
@@ -683,7 +683,7 @@ keywordsDF = addNewNames('https://raw.githubusercontent.com/pg-ufr-news/gensChuc
 keywordsDF = addNewNames('https://raw.githubusercontent.com/pg-ufr-news/winterIsComing/main/csv/sentiments_new_persons.csv', keywordsDF, 0.5, 'fr', 9)
 keywordsDF = addNewNames('https://raw.githubusercontent.com/pg-ufr-news/personWhisperer/main/csv/sentiments_new_persons.csv', keywordsDF, 0.5, 'fr', 9)
 '''
-
+keywordsDF['geonames'] = keywordsDF['geonames'].astype(int)
 keywordsDF = keywordsDF[(keywordsDF.ratioNew > 0.05)]
 #keywordsDF = keywordsDF.sort_values(by=['topic','keyword'])
 keywordsDF = keywordsDF.sort_values(by=['ratioNew'], ascending=False)
